[refactor] App: promote network models to domain models; change Endpoint to take Strings?
