+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/repository/mapper/ConsumptionMapper.kt
@@ -9,7 +9,7 @@ package com.rwmobi.kunigami.data.repository.mapper
 
 import com.rwmobi.kunigami.data.source.network.dto.ConsumptionDto
 import com.rwmobi.kunigami.domain.extensions.roundToNearestEvenHundredth
-import com.rwmobi.kunigami.domain.model.Consumption
+import com.rwmobi.kunigami.domain.model.consumption.Consumption
 
 fun ConsumptionDto.toConsumption() = Consumption(
     consumption = consumption.roundToNearestEvenHundredth(),
