+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/model/consumption/ConsumptionGrouping.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.model
+package com.rwmobi.kunigami.domain.model.consumption
 
 enum class ConsumptionGrouping(val apiValue: String?) {
     // This aggregates half hours into days based on the local time, not the UTC time, so there will be one fewer hour included on the daylight savings in the spring.
