+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/repository/RestApiRepository.kt
@@ -8,10 +8,12 @@
 package com.rwmobi.kunigami.domain.repository
 
 import com.rwmobi.kunigami.domain.model.Account
-import com.rwmobi.kunigami.domain.model.Consumption
 import com.rwmobi.kunigami.domain.model.Product
 import com.rwmobi.kunigami.domain.model.Rate
 import com.rwmobi.kunigami.domain.model.Tariff
+import com.rwmobi.kunigami.domain.model.consumption.Consumption
+import com.rwmobi.kunigami.domain.model.consumption.ConsumptionGrouping
+import com.rwmobi.kunigami.domain.model.consumption.ConsumptionOrdering
 import kotlinx.datetime.Instant
 
 interface RestApiRepository {
@@ -37,6 +39,8 @@ interface RestApiRepository {
         meterSerialNumber: String,
         periodFrom: Instant? = null,
         periodTo: Instant? = null,
+        orderBy: ConsumptionOrdering = ConsumptionOrdering.LATEST_FIRST,
+        groupBy: ConsumptionGrouping = ConsumptionGrouping.HALF_HOURLY,
     ): Result<List<Consumption>>
 
     suspend fun getAccount(apiKey: String, accountNumber: String): Result<List<Account>>
