+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/model/ConsumptionGroup.kt
@@ -8,7 +8,7 @@
 package com.rwmobi.kunigami.ui.model
 
 import androidx.compose.runtime.Immutable
-import com.rwmobi.kunigami.domain.model.Consumption
+import com.rwmobi.kunigami.domain.model.consumption.Consumption
 
 @Immutable
 data class ConsumptionGroup(
