+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/usecase/GetConsumptionUseCaseTest.kt
@@ -7,7 +7,7 @@
 
 package com.rwmobi.kunigami.domain.usecase
 
-import com.rwmobi.kunigami.domain.model.Consumption
+import com.rwmobi.kunigami.domain.model.consumption.Consumption
 import com.rwmobi.kunigami.domain.repository.FakeRestApiRepository
 import com.rwmobi.kunigami.domain.repository.FakeUserPreferencesRepository
 import io.kotest.matchers.shouldBe
